Added the ability to have multiple accessories (up to 3, min of 1)
The ability to have more than 1 accessory has been added, and the constructors have been updated too. 2 new constants have been added to the characterTraits class, one indicating the minimum number of accessories and the other one the max. A variable to indicate the number of accessories this character has has been added. The file reader/writer has been updated to support this new addition. The print properties was also updated to accept the new format.